+++ core/src/icbat/games/tradesong/engine/screens/PrototypeScreen.java
@@ -1,4 +1,4 @@
-package icbat.games.tradesong.engine;
+package icbat.games.tradesong.engine.screens;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.Screen;
@@ -6,6 +6,7 @@ import com.badlogic.gdx.graphics.GL20;
 import com.badlogic.gdx.scenes.scene2d.Stage;
 import com.badlogic.gdx.scenes.scene2d.ui.Table;
 import com.badlogic.gdx.utils.Timer;
+import icbat.games.tradesong.engine.PrototypeLayoutTable;
 import icbat.games.tradesong.game.Item;
 import icbat.games.tradesong.game.PlayerHoldings;
 import icbat.games.tradesong.game.TurnTaker;
